requirements.txt
aiofiles==0.8.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
asqlite==2.0.0a0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.7.0
asyncprawcore==2.3.0
attrs==22.2.0
braceexpand==0.1.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
import-expression==1.1.4
jishaku==2.5.1a472+gecdd201.master
multidict==6.0.4
psutil==5.9.4
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
tabulate==0.9.0
typing_extensions==4.5.0
update-checker==0.18.0
urllib3==1.26.15
yarl==1.8.2
